backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
fastapi==0.114.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
mangum==0.17.0
openai==1.46.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.1
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
